+++ moor_generator/lib/src/model/specified_column.dart
@@ -1,4 +1,5 @@
 import 'package:analyzer/dart/ast/ast.dart';
+import 'package:analyzer/dart/element/type.dart';
 import 'package:built_value/built_value.dart';
 
 part 'specified_column.g.dart';
@@ -99,9 +100,27 @@ class SpecifiedColumn {
   /// expression.
   final Expression defaultArgument;
 
+  /// If a type converter has been specified as the argument of
+  /// ColumnBuilder.map, this contains the Dart code that references that type
+  /// converter.
+  final Expression typeConverter;
+
+  /// If the type of this column has been overridden, contains the actual Dart
+  /// type. Otherwise null.
+  ///
+  /// Column types can be overridden with type converters. For instance, if
+  /// `C` was a type converter that converts `D` to `num`s, the column generated
+  /// by `real().map(const C())()` would have type `D` instead of `num`.
+  final DartType overriddenDartType;
+
   /// The dart type that matches the values of this column. For instance, if a
   /// table has declared an `IntColumn`, the matching dart type name would be [int].
-  String get dartTypeName => dartTypeNames[type];
+  String get dartTypeName {
+    if (overriddenDartType != null) {
+      return overriddenDartType.name;
+    }
+    return dartTypeNames[type];
+  }
 
   /// The column type from the dsl library. For instance, if a table has
   /// declared an `IntColumn`, the matching dsl column name would also be an
@@ -148,6 +167,8 @@ class SpecifiedColumn {
     this.nullable = false,
     this.features = const [],
     this.defaultArgument,
+    this.typeConverter,
+    this.overriddenDartType,
   });
 }